requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
astroid==3.0.2
    # via
    #   -r requirements.in
    #   pylint
bcrypt==4.1.2
    # via
    #   -r requirements.in
    #   flask-bcrypt
blinker==1.7.0
    # via
    #   -r requirements.in
    #   flask
build==1.0.3
    # via
    #   -r requirements.in
    #   pip-tools
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
click==8.1.7
    # via
    #   -r requirements.in
    #   flask
    #   pip-tools
cryptography==42.0.1
    # via -r requirements.in
dill==0.3.8
    # via
    #   -r requirements.in
    #   pylint
dnspython==2.5.0
    # via
    #   -r requirements.in
    #   email-validator
email-validator==2.1.0.post1
    # via -r requirements.in
flask==3.0.1
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-login
    #   flask-sqlalchemy
    #   flask-wtf
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
flask-wtf==1.2.1
    # via -r requirements.in
idna==3.6
    # via
    #   -r requirements.in
    #   email-validator
isort==5.13.2
    # via
    #   -r requirements.in
    #   pylint
itsdangerous==2.1.2
    # via
    #   -r requirements.in
    #   flask
    #   flask-wtf
jinja2==3.1.3
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.1.4
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
    #   wtforms
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   pylint
packaging==23.2
    # via
    #   -r requirements.in
    #   build
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   -r requirements.in
    #   pylint
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pylint==3.0.3
    # via -r requirements.in
pymysql==1.1.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
python-dotenv==1.0.1
    # via -r requirements.in
sqlalchemy==2.0.25
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
tomlkit==0.12.3
    # via
    #   -r requirements.in
    #   pylint
typing-extensions==4.9.0
    # via
    #   -r requirements.in
    #   sqlalchemy
werkzeug==3.0.1
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
wheel==0.42.0
    # via
    #   -r requirements.in
    #   pip-tools
wtforms==3.1.2
    # via
    #   -r requirements.in
    #   flask-wtf
